+++ client_src/js/dashboard.runner.js
@@ -13,9 +13,9 @@ import {
 } from "./dashboard/utils/chart.utils";
 import { DashUpdate } from "./dashboard/dashboard.updater";
 import {
-  dashboardSSEEventHandler,
+  dashboardSSEventHandler,
   workerURL
-} from "./dashboard/dashboard-sse.client";
+} from "./dashboard/dashboard-sse.handler";
 
 async function updateHistoryGraphs() {
   let historyStatistics = await OctoFarmclient.getHistoryStatistics();
@@ -72,7 +72,7 @@ async function initNewGraphs() {
   await ChartsManager.renderPrintCompletionByDay(printCompletionByDay);
 }
 
-createClientSSEWorker(workerURL, dashboardSSEEventHandler);
+createClientSSEWorker(workerURL, dashboardSSEventHandler);
 
 loadGrid()
   .then(async () => {