+++ client_src/js/monitoring/monitoring.updater.js
@@ -1,30 +1,23 @@
-import { dragAndDropEnable, dragCheck } from "../functions/dragAndDrop.js";
-import PrinterManager from "./printerManager.js";
-import PowerButton from "./powerButton.js";
-import UI from "../functions/ui.js";
-import Calc from "../functions/calc.js";
+import { dragAndDropEnable, dragCheck } from "../lib/functions/dragAndDrop.js";
+import PrinterManager from "../lib/modules/printerManager.js";
+import PowerButton from "../lib/modules/powerButton.js";
+import UI from "../lib/functions/ui.js";
+import Calc from "../lib/functions/calc.js";
 import {
   checkQuickConnectState,
   init as actionButtonInit
-} from "./Printers/actionButtons.js";
-import OctoPrintClient from "../octoprint.js";
-import { checkTemps } from "../modules/temperatureCheck.js";
-import { checkFilamentManager } from "./filamentGrab.js";
-import currentOperations from "./currentOperations";
-import doubleClickFullScreen from "../functions/fullscreen.js";
-import OctoFarmclient from "../octofarm";
+} from "../lib/modules/Printers/actionButtons.js";
+import OctoPrintClient from "../lib/octoprint.js";
+import { checkTemps } from "../lib/modules/temperatureCheck.js";
+import { checkFilamentManager } from "../lib/modules/filamentGrab.js";
+import doubleClickFullScreen from "../lib/functions/fullscreen.js";
+import OctoFarmclient from "../lib/octofarm";
+import { getControlList, getPrinterInfo } from "./monitoring-view.state";
 
 const elems = [];
 let powerTimer = 20000;
-let printerInfo = null;
-let printerControlList = null;
-let worker = null;
-let controlModal = false;
 let printerManagerModal = document.getElementById("printerManagerModal");
 let printerArea = document.getElementById("printerArea");
-let currentView = null;
-
-document.addEventListener("visibilitychange", handleVisibilityChange, false);
 
 document.getElementById("filterStates").addEventListener("change", (e) => {
   OctoFarmclient.get("client/updateFilter/" + e.target.value);
@@ -33,58 +26,17 @@ document.getElementById("sortStates").addEventListener("change", (e) => {
   OctoFarmclient.get("client/updateSorting/" + e.target.value);
 });
 
-function handleVisibilityChange() {
-  if (document.hidden) {
-    if (worker !== null) {
-      console.log("Screen Abandonded, closing web worker...");
-      worker.terminate();
-      worker = null;
-    }
-  } else {
-    if (worker === null) {
-      console.log("Screen resumed... opening web worker...");
-      createWebWorker(currentView);
-    }
-  }
-}
 const returnPrinterInfo = (id) => {
+  const statePrinterInfo = getPrinterInfo();
   if (typeof id !== "undefined") {
-    const zeeIndex = _.findIndex(printerInfo, function (o) {
+    const zeeIndex = _.findIndex(statePrinterInfo, function (o) {
       return o._id == id;
     });
-    return printerInfo[zeeIndex];
+    return statePrinterInfo[zeeIndex];
   } else {
-    return printerInfo;
+    return statePrinterInfo;
   }
 };
-export default function createWebWorker(view) {
-  currentView = view;
-  worker = new Worker("/assets/dist/monitoringViewsWorker.min.js");
-  worker.onmessage = async function (event) {
-    if (event.data != false) {
-      //Update global variables with latest information...
-      printerInfo = event.data.printersInformation;
-      printerControlList = event.data.printerControlList;
-      //Grab control modal element...
-      if (!controlModal) {
-        controlModal = document.getElementById("printerManagerModal");
-      }
-      await init(
-        event.data.printersInformation,
-        event.data.clientSettings,
-        currentView
-      );
-      if (event.data.clientSettings.panelView.currentOp) {
-        const currentOperationsData = event.data.currentOperations;
-        currentOperations(
-          currentOperationsData.operations,
-          currentOperationsData.count,
-          printerInfo
-        );
-      }
-    }
-  };
-}
 
 function isHidden(state, clientSettings) {
   let hidden = "";
@@ -93,11 +45,9 @@ function isHidden(state, clientSettings) {
   } else if (state === "Disconnected" && clientSettings.panel.hideClosed) {
     hidden = "hidden";
   }
-  // } else if (state === "Idle" && clientSettings.hideIdle) {
-  //   hidden = "hidden";
-  // }
   return hidden;
 }
+
 function isRotated(otherSettings) {
   let flipH = "";
   let flipV = "";
@@ -116,6 +66,7 @@ function isRotated(otherSettings) {
   }
   return { flipH, flipV, rotate90 };
 }
+
 function cleanName(printerName) {
   let name = printerName;
   if (name.includes("http://")) {
@@ -137,6 +88,7 @@ function checkPrinterRows(clientSettings) {
     return 2;
   }
 }
+
 function imageOrCamera(printer) {
   let drawCamera = ({ url, flipV, flipH, rotate90 }) => {
     return `<img
@@ -150,6 +102,7 @@ function imageOrCamera(printer) {
   };
   const flip = isRotated(printer.otherSettings);
   const { flipH, flipV, rotate90 } = flip;
+
   //Is octoprints camera settings enabled?
   if (
     printer.otherSettings !== null &&
@@ -302,6 +255,7 @@ function drawListView(printer, clientSettings) {
         </tr>
     `;
 }
+
 function drawPanelView(printer, clientSettings) {
   const hidden = isHidden(printer, clientSettings);
   const name = cleanName(printer.printerName);
@@ -463,6 +417,7 @@ function drawPanelView(printer, clientSettings) {
       </div>
     `;
 }
+
 function drawCameraView(printer, clientSettings) {
   let hidden = isHidden(printer, clientSettings);
   if (printer.cameraURL === "") {
@@ -530,27 +485,17 @@ function drawCameraView(printer, clientSettings) {
           
           <div class="camExtra">
             <div class=" row">
-
               <div class="col-6">
-                 <span
-                    class="mb-0 text-center"
-                    id="printTimeElapsed-${printer._id}"
-                  >
-
-                  </span>
+                 <span class="mb-0 text-center" id="printTimeElapsed-${printer._id}"></span>
               </div>
               <div class="col-6">
-                          <span
-              class="mb-0 text-center"
-              id="remainingTime-${printer._id}"
-            >
-
-            </span>
+                <span class="mb-0 text-center" id="remainingTime-${printer._id}"></span>
               </div>
-                          
-</div>
+            </div>
           </div>
-                ${cameraElement}
+          
+          ${cameraElement}
+          
           <div class="camTemps">
             <small
               id="toolTemps-${printer._id}"
@@ -600,9 +545,10 @@ function addListeners(printer) {
   //For now Control has to be seperated
   document
     .getElementById(`printerButton-${printer._id}`)
-    .addEventListener("click", () => {
-      // eslint-disable-next-line no-underscore-dangle
-      PrinterManager.init(printer._id, printerInfo, printerControlList);
+    .addEventListener("click", async () => {
+      const printerInfo = getPrinterInfo();
+      const controlList = getControlList();
+      await PrinterManager.init(printer._id, printerInfo, controlList);
     });
 
   //Play button listeners
@@ -1259,12 +1205,12 @@ async function updateState(printer, clientSettings, view) {
   }
 }
 
-async function init(printers, clientSettings, view) {
+export async function initMonitoring(printers, clientSettings, view) {
   //Check if printer manager modal is opened
   switch (printerManagerModal.classList.contains("show")) {
     case true:
       // Run printer manager updater
-      PrinterManager.init("", printers, printerControlList);
+      await PrinterManager.init("", printers, clientSettings);
       break;
     case false:
       // initialise or start the information updating..