+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
     - Refactor of Printer Manager client code bringing a little speed boost (avg 40fps -> 50fps) and better fault tolerance. 
     - Refactor of SSE client into re-usable file. 
     - Moved the "API Errors" button, now appears on the start of a printer row and clicking will take you straight into Printer Settings.
+    - Refactor monitoring pages (panel, list, camera) with latest SSE client reusability.
 
 ### Removed